week-02/grpc-streaming/client/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.1